Unhook 'hunt.join' from Roles
