+++ imports/lib/roles.ts
@@ -2,7 +2,6 @@ import { Roles } from 'meteor/nicolaslopezj:roles';
 
 Roles.registerAction('gdrive.credential', true);
 Roles.registerAction('google.configureOAuth', true);
-Roles.registerAction('hunt.join', true);
 Roles.registerAction('hunt.bulkJoin', true);
 Roles.registerAction('discord.configureOAuth', true);
 Roles.registerAction('discord.configureBot', true);
